requirements.txt
# Core
torch>=2.2.0
torchvision>=0.17.0
torchaudio>=2.2.0

# Imágenes
Pillow>=10.0.0
opencv-python>=4.8.0

# Análisis y visualización
numpy>=1.26.0
matplotlib>=3.8.0
seaborn>=0.13.0
scikit-learn>=1.4.0

# Notebooks
jupyterlab>=4.0.0
ipywidgets>=8.0.0

# Utilidades
tqdm>=4.66.0
pandas>=2.1.0

# API (opcional)
fastapi>=0.110.0
uvicorn>=0.27.0
python-multipart>=0.0.9

# Tracking (opcional)
# wandb>=0.16.0

# Arquitecturas adicionales (opcional)
# timm>=0.9.0
